fix for kotlin.java.stdlib.jar problem

+++ components-registry-service-server/src/main/kotlin/org/octopusden/octopus/components/registry/server/listener/StartupApplicationListener.kt
@@ -36,7 +36,10 @@ class StartupApplicationListener: ApplicationListener<ApplicationStartingEvent> 
 
     override fun onApplicationEvent(event: ApplicationStartingEvent) {
         val dslKotlinModule = Thread.currentThread().contextClassLoader.getResource("/components-registry-dsl.txt")
-        if (dslKotlinModule != null && dslKotlinModule.toString().matches(Regex(".*!/BOOT-INF/.*"))) {
+        println("DEBUG: StartupApplicationListener running. dslKotlinModule=$dslKotlinModule")
+        
+        if (dslKotlinModule != null && dslKotlinModule.toString().contains("!BOOT-INF/")) {
+            println("DEBUG: Spring boot jar running mode detected")
             LOG.debug("Spring boot jar running mode detected")
             temporaryLibraryPath = Files.createTempDirectory("components-registry-dsl-" + UUID.randomUUID())
             val dslLibraryClassPath = StringBuffer()
@@ -45,13 +48,29 @@ class StartupApplicationListener: ApplicationListener<ApplicationStartingEvent> 
                 Files.walk(fs.getPath("/BOOT-INF/lib")).use { filePath ->
                     libraryFiles.addAll(filePath.filter { it.toString().endsWith(".jar") }.toList())
                 }
+                
+                // Fix for kotlin.java.stdlib.jar property issue in fat jar
+                val stdlibJar = libraryFiles.find { it.fileName.toString().matches(Regex("kotlin-stdlib-[0-9].*\\.jar")) }
+                if (stdlibJar != null) {
+                    val dstPath = temporaryLibraryPath!!.resolve(stdlibJar.fileName.toString())
+                    if (!Files.exists(dstPath)) {
+                        Files.copy(stdlibJar, dstPath)
+                    }
+                    System.setProperty("kotlin.java.stdlib.jar", dstPath.toString())
+                    LOG.info("Set kotlin.java.stdlib.jar to $dstPath")
+                } else {
+                    LOG.warn("Unable to find kotlin-stdlib jar in BOOT-INF/lib")
+                }
+
                 dslKotlinModule.openStream().use {inputStream ->
                     inputStream.reader().forEachLine { fileName ->
                         val libraryName = fileName.split(LIBRARY_VERSION_SPLIT_REGEXP, 2)[0]
                         val srcPath = libraryFiles.findLast { it.fileName.toString().split(LIBRARY_VERSION_SPLIT_REGEXP, 2)[0] == libraryName }
                                 ?: throw IllegalStateException("Unable to match provided library $fileName")
                         val dstPath = temporaryLibraryPath!!.resolve(srcPath.fileName.toString())
-                        Files.copy(srcPath, dstPath)
+                        if (!Files.exists(dstPath)) {
+                            Files.copy(srcPath, dstPath)
+                        }
                         dslLibraryClassPath.append(File.pathSeparator).append(dstPath)
                     }
                 }
